mqtt-client/requirements.txt
cryptography==1.9
pyjwt==1.5.2
paho-mqtt==1.3.0
requests